+++ formula-android-tests/src/main/java/com/instacart/formula/test/TestContract.kt
@@ -1,14 +1,14 @@
-package com.instacart.formula
+package com.instacart.formula.test
 
 import android.view.View
+import com.instacart.formula.R
 import com.instacart.formula.fragment.FragmentComponent
 import com.instacart.formula.fragment.FragmentContract
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
-data class TestContractWithId(
-    val id: Int,
-    override val tag: String = "test-contract-$id",
+data class TestContract(
+    override val tag: String = "test contract",
     override val layoutId: Int = R.layout.test_empty_layout
 ) : FragmentContract<Any>() {
     override fun createComponent(view: View): FragmentComponent<Any> {
