+++ formula-android-tests/src/main/java/com/instacart/formula/test/TestFragmentActivity.kt
@@ -1,7 +1,8 @@
-package com.instacart.formula
+package com.instacart.formula.test
 
 import android.os.Bundle
 import androidx.annotation.VisibleForTesting
+import com.instacart.formula.R
 import com.instacart.formula.fragment.FormulaFragment
 import com.instacart.formula.fragment.FragmentContract
 import com.instacart.formula.integration.FormulaAppCompatActivity
