+++ jacoco.sh
@@ -1,0 +1,2 @@
+./test.sh
+./gradlew jacocoTestReportMerged --stacktrace
